+++ Sources/swift-evolution-metadata-extractor/SnapshotCommand.swift
@@ -28,10 +28,13 @@ struct SnapshotCommand: AsyncParsableCommand {
     @Flag(name: .shortAndLong, help: Help.Shared.Argument.verbose)
     var verbose: Bool = false
 
+    @Argument(help: Help.Shared.Argument.proposalFiles, transform: ArgumentValidation.proposalURL)
+    var proposalURLs: [URL]
+
     mutating func validate() throws {
         ArgumentValidation.validate(verbose: verbose)
         ArgumentValidation.validateHTTPProxies()
-        extractionSource = try ArgumentValidation.extractionSource(snapshotURL: snapshotURL)
+        extractionSource = try ArgumentValidation.extractionSource(snapshotURL: snapshotURL, proposalURLs: proposalURLs)
     }